+++ plugins/admission/pkg/models/models.go
@@ -37,3 +37,11 @@ type Configuration struct {
 	}
 	Polaris *polarisconfiguration.Configuration
 }
+
+type InsightsConfig struct {
+	Hostname        string
+	Organization    string
+	Cluster         string
+	Token           string
+	IgnoreUsernames []string
+}