+++ plugins/workloads/pkg/info.go
@@ -139,21 +139,36 @@ func CreateResourceProviderFromAPI(ctx context.Context, dynamicClient dynamic.In
 		topController := workload.TopController
 		var containers []ContainerResult
 		podCount := float64(len(workload.Pods))
-		pod := workload.Pods[0]
-		// Convert the unstructured object to cluster.
+
 		var pd corev1.Pod
-		err = runtime.DefaultUnstructuredConverter.
-			FromUnstructured(pod.UnstructuredContent(), &pd)
-		if err != nil {
-			return nil, err
+
+		podSpec := controller.GetPodSpec(workload.TopController.Object)
+		if podSpec == nil {
+			// Could be a top-level object like Prometheus, which doesn't have a podSpec, so fall back to the actual pods
+			if len(workload.Pods) > 0 {
+				err = runtime.DefaultUnstructuredConverter.FromUnstructured(workload.Pods[0].UnstructuredContent(), &pd)
+				if err != nil {
+					return nil, err
+				}
+			} else {
+				// Nothing we can do here--no pods in the cluster, and no podSpec in the top-level object
+				// TODO: there's probably a mid-level object where we can get the info.
+				// e.g. a Prometheus doesn't have podSpec, but its Deployment does
+				continue
+			}
+		} else {
+			err = runtime.DefaultUnstructuredConverter.FromUnstructured(map[string]interface{}{"spec": podSpec}, &pd)
+			if err != nil {
+				return nil, err
+			}
 		}
+		// Convert the unstructured object to cluster.
 		for _, ctn := range pd.Spec.Containers {
 			containers = append(containers, formatContainer(ctn, corev1.ContainerStatus{}, topController.GetCreationTimestamp()))
 		}
 		controller := formatControllers(topController.GetKind(), topController.GetName(), topController.GetNamespace(), string(topController.GetUID()), topController.GetOwnerReferences(), containers, topController.GetAnnotations(), topController.GetLabels())
 		controller.PodCount = podCount
 		interfaces = append(interfaces, controller)
-
 	}
 
 	// Nodes